PR comments fixed, deleted transferTokenBuilder test file, moved those tests to transferBuilder file.
Ticket: BGA-13
